20220605_暂时屏蔽-Werror
(1)类型转换问题
(2)switch 分支尾端没有写 break;
(3)using namespace XXX 的问题;
这些都被目前的 -Werror 强制判为 error...

+++ src/Engine/map/MapEnt.h
@@ -49,7 +49,7 @@ public:
         density(density_),
         mapAlti(mapAlti_),
         bioSoupState(bioSoupState_),
-        uWeight(uWeight_),
+        uWeight( static_cast<size_t>(uWeight_) ),
         isEcoBorder(isEcoBorder_)
         {}
